Rename init-utils as commercial-boot-utils
