+++ src/init/consented-advertising.ts
@@ -1,5 +1,5 @@
 import { init as prepareAdVerification } from '../lib/ad-verification/prepare-ad-verification';
-import { bootCommercial } from '../lib/init-utils';
+import { bootCommercial } from '../lib/commercial-boot-utils';
 import { adFreeSlotRemove } from './consented/ad-free-slot-remove';
 import { init as initComscore } from './consented/comscore';
 import { initDfpListeners } from './consented/dfp-listeners';
